Move lyrics related functions to lyrics.py
