+++ src/player/player.py
@@ -7,7 +7,6 @@ from re import match
 
 from discord import Embed, FFmpegPCMAudio
 from discord.ext.commands import Context
-from src.player.genius import GeniusApi
 from src.player.songcache import SongCache
 from src.player.youtube import (download_song, get_song_url,
                                 get_youtube_playlist_songlist)
@@ -21,7 +20,6 @@ class Player():
         self.logger = bot.logger
         self.cache = SongCache(self.logger)
         self.IDLE_TIMEOUT = getenv("IDLE_TIMEOUT", 1)
-        self.genius = GeniusApi(bot.logger)
         self.playing = False
 
     async def play(self, ctx: Context, play_text: str) -> None:
@@ -73,8 +71,7 @@ class Player():
         user_voice_channel = ctx.author.voice.channel
         if ctx.voice_client and ctx.voice_client.channel == user_voice_channel:
             queue = self.get_queue(ctx)
-            await self.bot.send_commands_list(ctx)
-            await ctx.voice_client.disconnect()
+            await self.bot.leave(ctx)
             with queue.mutex:
                 queue.queue.clear()
             self.playing = False
@@ -115,7 +112,7 @@ class Player():
                 self.playing = True
                 self.current_song[ctx.guild.id] = queue.get()
 
-                await self.bot.clear_bot_msgs_in_channel(self.bot.user, ctx)
+                await self.bot.clear_bot_msgs_in_channel(ctx)
 
                 current_playing_song_embed_msg = Embed(title=f":arrow_forward: **Reproduzindo**",
                                                        description=f"`{self.current_song[ctx.guild.id].title}`", color=0x550a8a)
@@ -144,8 +141,7 @@ class Player():
 
             await sleep(1)
             idle_timer += 1
-        await self.bot.send_commands_list(ctx)
-        await voice_client.disconnect()
+        await self.bot.leave(ctx)
         self.logger.info('O bot desconectou do canal após reproduzir a fila.')
         return
 
@@ -215,20 +211,6 @@ class Player():
                     ctx.message.channel.send(embed=embed_msg))
                 self.logger.info(f'O bot embaralhou a fila.')
 
-    async def lyrics(self, ctx: Context, search_text: str = None) -> None:
-        """
-        Send lyrics from the current song or from a search text using Genius API.
-        """
-
-        if search_text is None:
-            self.bot.loop.create_task(
-                self.send_current_song_lyrics(ctx)
-            )
-        else:
-            self.bot.loop.create_task(
-                self.send_lyrics_by_search_text(ctx, search_text=search_text)
-            )
-
     async def handle_song_request(self, play_text: str, ctx: Context,) -> None:
         is_youtube_playlist = match(
             "https://www.youtube.com/playlist*", play_text)
@@ -240,40 +222,6 @@ class Player():
             await self.add_song(play_text, ctx, link=True)
         else:
             await self.add_song(play_text, ctx)
-
-    async def send_current_song_lyrics(self, ctx: Context) -> None:
-        """
-        Send lyrics from the current song.
-        """
-        current_song = self.current_song[ctx.guild.id]
-        await self.send_lyrics_by_search_text(ctx, current_song.title)
-
-    async def send_lyrics_by_search_text(self, ctx: Context, search_text: str = None) -> None:
-        """
-        Send lyrics by search text.
-        """
-        searching_embed_msg = Embed(title=f":mag_right: **Procurando letra da música**: `{search_text}`",
-                                    color=0x550a8a)
-        msg = await ctx.send(embed=searching_embed_msg)
-
-        song = await self.genius.get_song_with_lyrics(search_text)
-        if msg:
-            await msg.delete()
-
-        lyrics_embed_msg = ''
-        if song and song.lyrics:
-            self.logger.info(
-                f'O bot enviou a lyrics ao canal')
-
-            lyrics_embed_msg = Embed(title=f":pencil: **Lyrics**",
-                                     description=f"**{song.title} by {song.artist}**\n\n{(song.lyrics)}",
-                                     color=0x550a8a)
-        else:
-            self.logger.info(f'O bot não encontrou a lyrics.')
-
-            lyrics_embed_msg = Embed(title=f":x: **Lyrics não encontrada**",
-                                     color=0xeb2828)
-        await ctx.message.channel.send(embed=lyrics_embed_msg)
 
     async def add_playlist(self, play_list_url: str, ctx: Context) -> None:
         """
